refactor(cc-dispatcher): extract WORKFLOW_END_USER_MESSAGES to cc-workflow-end-messages.ts
Pure-data extraction per #3243 status comment (PR #3802). Moves the
~15-LoC user-copy map + compile-time exhaustiveness rail out of
cc-dispatcher.ts into a new sibling module. No behavior change; the
two consumer reads at cc-dispatcher.ts:1626/1634 are byte-identical
(line numbers shifted -36 by the deletion).

Type-source uses ./soleur-go-runner's WorkflowEnd["status"] (mirrors
cc-dispatcher.ts:212), NOT @/lib/types' WorkflowEndStatus — the
lib/types enum has 9 wire-protocol values while the runner emits 7;
using lib/types would fire the exhaustiveness rail on the missing
sandbox_denial + runner_crash keys. This drift is pre-existing and
filed as a follow-up scope-out (see ADR-031 + plan Risks).

Test block at test/cc-dispatcher.test.ts:723-769 relocated to new
test/cc-workflow-end-messages.test.ts with static import (matches
test/cc-cost-caps.test.ts:9 sibling-extraction precedent).

Full apps/web-platform vitest suite green: 4367 passed / 39 skipped.
Typecheck clean.

Ref #3243
