+++ knowledge-base/project/specs/feat-one-shot-cc-dispatcher-extract-workflow-end-messages/session-state.md
@@ -1,0 +1,21 @@
+# Session State
+
+## Plan Phase
+- Plan file: knowledge-base/project/plans/2026-05-15-refactor-extract-cc-workflow-end-messages-plan.md
+- Status: complete
+
+### Errors
+None. Two non-blocking warnings: Bash CWD persistence reset between calls (mitigated via absolute paths); harness TaskCreate reminders for a two-step skill-invocation task (ignored).
+
+### Decisions
+- Type-source pivot: import `WorkflowEnd` from `./soleur-go-runner`, derive `WorkflowEndStatus = WorkflowEnd["status"]` locally — `lib/types.ts` has 9 statuses vs runner's 7, so the names drifted; using runner type matches cc-dispatcher.ts:212 precedent.
+- Move existing `WORKFLOW_END_USER_MESSAGES` test block from test/cc-dispatcher.test.ts:723-769 into a new test/cc-workflow-end-messages.test.ts with static import (matches cc-cost-caps.test.ts:9 precedent).
+- Data-only module: no functions, no logger, no side effects; strictly fewer deps than source site. cc-dispatcher.ts:212 local `WorkflowEndStatus` re-derive stays (out of scope — still used by TERMINAL_WORKFLOW_END_STATUSES, ABORT_FLUSH_STATUSES, AbortFlushStatus).
+- PR body uses `Ref #3243` (not `Closes`) — multiple extractions still pending; next-next is `cc-singletons.ts` (reaper + rate-limiter).
+- Deepen calibrated to MORE (not A LOT) — ~50-LoC pure-data extraction; load-bearing finding was the type-source drift caught by direct code reads.
+
+### Components Invoked
+- skill: soleur:plan (revalidation triad: wc -l, mirrorWithDebounce import-vs-define check, WORKFLOW_END_USER_MESSAGES purity check; ADR-030 + cc-cost-caps precedent studied).
+- skill: soleur:deepen-plan (User-Brand Impact halt gate passed; type-source verification via Read on lib/types.ts and soleur-go-runner.ts; gh pr/issue/label live-verification).
+- Plan-time bash recon: wc -l, git grep, gh API verification of PRs 3608/3670/3802 and issue 3243.
+- No external review-agent fan-out (deliberate cost calibration for aggregate-pattern threshold refactor).
